+++ editor/app/(api)/private/workspace/[organization_id]/projects/[project_id]/tags/route.ts
@@ -1,0 +1,41 @@
+import { createRouteHandlerWorkspaceClient } from "@/lib/supabase/server";
+import { cookies } from "next/headers";
+import { notFound } from "next/navigation";
+import { type NextRequest, NextResponse } from "next/server";
+import type { Platform } from "@/lib/platform";
+
+type Params = { organization_id: number; project_id: number };
+
+export async function GET(
+  req: NextRequest,
+  context: { params: Promise<Params> }
+) {
+  const { project_id } = await context.params;
+  const cookieStore = await cookies();
+  const client = createRouteHandlerWorkspaceClient(cookieStore);
+  //
+
+  const { data: tags, error: projects_err } = await client
+    .from("tag")
+    .select(
+      `
+      *,
+      customer_tag(count)
+    `
+    )
+    .eq("project_id", project_id);
+
+  if (!tags) {
+    return notFound();
+  }
+
+  const tagsWithUsage = tags?.map((tag) => ({
+    ...tag,
+    usage_count: tag.customer_tag[0].count,
+  })) satisfies Platform.Tag.TagWithUsageCount[];
+
+  return NextResponse.json({
+    data: tagsWithUsage,
+    error: null,
+  });
+}
