+++ editor/app/(workbench)/[org]/[proj]/(project)/layout.tsx
@@ -16,7 +16,7 @@ import { EditorHelpFab } from "@/scaffolds/help/editor-help-fab";
 import { Inter } from "next/font/google";
 import Link from "next/link";
 import { ArrowLeftIcon } from "@radix-ui/react-icons";
-import { ProjectLoaded } from "@/scaffolds/workspace";
+import { ProjectLoaded, ProjectTagsProvider } from "@/scaffolds/workspace";
 import { ResourceTypeIcon } from "@/components/resource-type-icon";
 import { AboutGridaWestCard } from "./about-west-card";
 import "../../../../editor.css";
@@ -130,7 +130,9 @@ export default async function Layout({
                   <div className="flex flex-col overflow-hidden w-full h-full">
                     <header className="px-2 h-11 min-h-11 flex items-center border-b bg-workbench-panel desktop-drag-area"></header>
                     <div className="w-full h-full overflow-x-hidden overflow-y-auto">
-                      <ProjectLoaded>{children}</ProjectLoaded>
+                      <ProjectLoaded>
+                        <ProjectTagsProvider>{children}</ProjectTagsProvider>
+                      </ProjectLoaded>
                     </div>
                   </div>
                 </div>
