+++ editor/app/(workbench)/[org]/[proj]/(project)/tags/tag-form-dialog.tsx
@@ -24,9 +24,9 @@ import {
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import { Textarea } from "@/components/ui/textarea";
+import { useTags } from "@/scaffolds/workspace";
+import type { Platform } from "@/lib/platform";
 import toast from "react-hot-toast";
-import { Platform } from "@/lib/platform";
-import { useTags } from "./context";
 
 const tagFormSchema = z.object({
   name: z
